Hopefully fix builds after rename
