+++ src/frontends/widgets/views/economy_view.cpp
@@ -17,8 +17,8 @@
 
 #include "economy_view.h"
 
-#include "../../../model/empire.h"
-#include "../../../model/galaxy_state.h"
+#include "../../../core/empire.h"
+#include "../../../core/galaxy_state.h"
 #include "../numerictableitem.h"
 
 EconomyView::EconomyView(QWidget *parent) : QTableWidget(parent) {
