+++ src/frontends/widgets/views/fleets_view.cpp
@@ -17,11 +17,11 @@
 
 #include "fleets_view.h"
 
-#include "../../../model/empire.h"
-#include "../../../model/fleet.h"
-#include "../../../model/galaxy_state.h"
-#include "../../../model/ship.h"
-#include "../../../model/ship_design.h"
+#include "../../../core/empire.h"
+#include "../../../core/fleet.h"
+#include "../../../core/galaxy_state.h"
+#include "../../../core/ship.h"
+#include "../../../core/ship_design.h"
 #include "../numerictableitem.h"
 
 FleetsView::FleetsView(QWidget *parent) : QTableWidget(parent) {
