+++ src/frontends/widgets/views/overview_view.cpp
@@ -17,8 +17,8 @@
 
 #include "overview_view.h"
 
-#include "../../../model/galaxy_state.h"
-#include "../../../model/empire.h"
+#include "../../../core/galaxy_state.h"
+#include "../../../core/empire.h"
 #include "../numerictableitem.h"
 
 OverviewView::OverviewView(QWidget *parent) : QTableWidget(parent) {
